"Added stock management and cart item removal functionality to cart and product components"
This commit includes changes to both the Cart.jsx and Product.jsx components, as well as the addition of a new CartItem.jsx component. The changes allow for the display and management of product stock levels, as well as the removal of items from the cart. The CartItem component includes a new removeProduct function that uses axios to send a DELETE request to the server to remove a product from the cart. The Product component has been updated to display whether a product is in stock or not, and the Cart component now includes a remove button for each cart item. Additionally, the cart.css and product.css files have been updated to include new styles for the added functionality.